dodałem hp do gracza

+++ src/object.h
@@ -95,7 +95,7 @@ namespace Entity {
         std::array<std::shared_ptr<Items::AbstractItem>, STACK_SIZE> items;
         int selectedItem;
 
-        int hp;
+        float hp;
 
         public:
         Player(Vector2 pos);
@@ -110,8 +110,8 @@ namespace Entity {
         void setLookingDirection(Vector2);
         Vector2 getLookingDirection() const;
 
-        void setHp(int nowe);
-        int getHp();
+        void setHp(float nowe);
+        float getHp();
     };
 
     class DropedItem: public PhysicsObject {
